Implement basic controls
This starts by discovering the host and port the HTTP API is bound on by
talking to GoXLR-Utility through its socket. Once that's completed, it
opens up a websocket connection, retrieves the initial state of the
mixer(s) and then applies any state changes to it.

The old/initial and new state are passed together over a channel. Each
time a message is received, all the Expr scripts are executed with the
previous and new state passed to them.

Functionality to send the PlayPause event over D-Bus for your music
player of choice is included.