Add ability to execute commands
Using RunCommand you can execute any binary on a system with whatever
arguments. Commands are constrained to running for up to 1sec.

Commands will be looked up in path, unless an absolute path is given.

This does not execute a shell and so cannot directly execute shell
scripts. For that, you can call it with `sh -c`.