Remove from ConditionIfPathExists from unit file
The ConditionIfPathExists seems to get evaluated very early, before
goxlr-daemon has had a chance to start, or at least before it's had a
chance to create the socket file. This results in GopherXLR never
starting.

Since it's expected to start GoXLR-Utility as part of a user session
using autostart, this switches the dependency on the generated autostart
service instead.