UnpackResponse no longer hangs indefinitely
GoXLR-Utility 1.1.x includes a fix that ensures that if an invalid
command is sent to the daemon using the IPC mechanism, an error response
is now returned. This means that we should never hang indefinitely in
UnpackResponse.

It's still good practice to manage the read deadline so MustGetAddress
continues to do that, as there are always other ways things can fail.